requirements.txt
pandas
requests
beautifulsoup4
selenium
pillow
